Refactor for common vim bundle install
I'm not 100% happy with how it looks/works, and feel like some of
the bundle information should go into a hash somewhere, then I can
just pass an array of bundle hashes, and it would be very clean.
Save that for later, if ever, because this already seems a bit
overcomplicated to me.