Added basic gitconfig to bootstrap setup
I know that this is supposed to be a first-run type setup on a new machine, but I
often run this locally to test any changes. If there was some way to collect/collate
existing prefs as well as new prefs (for git or whatever else) that'd be cool.
But I guess that's outside the scope/intent of a dotfile bootstrap setup.